Eduman/requirements.txt
streamlit
Pillow
google-generativeai
protobuf
python-dotenv
